Add node:readme subgenerator
This merge node:app and node:package as the later one is at the base of
everything else we'll generate here. Let's keep it top level.
